+++ life_expectancy/cleaning.py
@@ -1,46 +1,4 @@
-from inspect import getsourcefile
-from os.path import dirname, abspath, join
-import argparse
 import pandas as pd
-
-
-def get_current_directory_full_path() -> str:
-    """
-    Returns the absolute path of the directory containing the current 
-    Python file.
-    :return: A string representing the absolute path of the directory 
-    containing the current Python file.
-    """
-    current_file_path: str = abspath(getsourcefile(lambda: 0))
-    current_dir_path: str = dirname(current_file_path)
-
-    return current_dir_path
-
-
-def load_data(
-        path: str = None
-) -> pd.DataFrame:
-    """
-    Loads the 'eu_life_expectancy_raw.tsv' data file from the 'data' folder
-    :param path: path to file to load.
-    :return: Pandas dataframe with data
-    :raises FileNotFoundError: If the 'eu_life_expectancy_raw.tsv' file cannot
-     be found.
-    """
-    # Ensure path
-    if path is None:
-        path = join(
-            get_current_directory_full_path(),
-            "data",
-            "eu_life_expectancy_raw.tsv"
-        )
-    # Load data
-    raw_data: pd.DataFrame = pd.read_csv(
-        path,
-        sep="\t",
-        header=0
-    )
-    return raw_data
 
 
 def clean_data(
@@ -117,51 +75,3 @@ def clean_data(
     long_data = long_data.dropna(subset=["value"])
 
     return long_data
-
-
-def save_data(data: pd.DataFrame) -> None:
-    """
-    Data is saved to a CSV file named 'pt_life_expectancy.csv' in the 'data'
-    folder.
-    :param data: 
-    :return: 
-    """
-    data.to_csv(
-        join(
-            get_current_directory_full_path(),
-            "data",
-            "pt_life_expectancy.csv"
-        ),
-        index=False
-    )
-
-
-def main(
-        country_code: str = 'PT',
-        path: str = None
-) -> pd.DataFrame:
-    """
-    Calls load data, clean data and save data
-    :param country_code: A string representing the country code of the country
-    to be selected.
-    :param path: path to file to load.
-    :return: Cleaned data frame
-    """
-    wide_data = load_data(path=path)
-    cleaned_data = clean_data(
-        wide_data=wide_data,
-        country_code=country_code)
-    save_data(data=cleaned_data)
-    return cleaned_data
-
-
-if __name__ == "__main__":  # pragma: no cover
-    parser = argparse.ArgumentParser(description='Clean life expectancy data.')
-    parser.add_argument(
-        '--country',
-        type=str, default='PT',
-        help='Country code to filter data'
-    )
-    args = parser.parse_args()
-
-    main(country_code=args.country)
