+++ life_expectancy/tests/test_cleaning.py
@@ -1,16 +1,21 @@
 """Tests for the cleaning module"""
 import pandas as pd
-
-from life_expectancy.cleaning import main
-from . import OUTPUT_DIR, FIXTURES_DIR
+from life_expectancy.cleaning import clean_data
 
 
-def test_main(pt_life_expectancy_expected, life_expectancy_data):
-    """Run the `main` function and compare the output to the expected output"""
-    main(path=FIXTURES_DIR / "eu_life_expectancy_input.tsv")
-    pt_life_expectancy_actual = pd.read_csv(
-        OUTPUT_DIR / "pt_life_expectancy.csv"
+def test_clean(
+        pt_life_expectancy_expected,
+        life_expectancy_data
+):
+    """
+    Test clean function
+    :param pt_life_expectancy_expected:
+    :param life_expectancy_data:
+    :return:
+    """
+    cleaned_data = clean_data(
+        wide_data=life_expectancy_data
     )
     pd.testing.assert_frame_equal(
-        pt_life_expectancy_actual, pt_life_expectancy_expected
+        cleaned_data.reset_index(drop=True), pt_life_expectancy_expected
     )
